test: split private PDF fidelity discovery cases
