+++ tools/pdf-private-fidelity-boundary.cases.mjs
@@ -1,0 +1,1473 @@
+import { describe, expect, it } from 'vitest'
+import {
+  chmodSync, mkdirSync, mkdtempSync, readFileSync, readdirSync, rmSync, statSync, symlinkSync, writeFileSync, homedir, tmpdir, join, resolve, spawnSync, createHash, fileURLToPath, canonicalHyphenEvidenceSha256, canonicalJsonHash, PDF_HYPHEN_DERIVED_AFFIX_REMOVAL_REQUIRED_EVIDENCE, PDF_HYPHEN_LEXICAL_MODEL_RECEIPT, PDF_HYPHEN_PRODUCTIVE_PREFIX_RULE_RECEIPT, PDF_HYPHEN_REMOVAL_REQUIRED_EVIDENCE, applyPrivateDecisionSet, comparePrivateFidelityBaseline, comparePrivateFidelityReceipts, createPrivateFidelityReceipt, createPrivateFidelityRunReceipt, createPrivateReconstructionEvidenceRaw, parsePrivateFidelityArguments, prepareOwnerOnlyDirectory, writeExclusive, privateFidelity, hash, privateCompletenessPolicy, createPrivateReconstructionEvidence, canonicalHyphenDeletionRecord, canonicalDerivedAffixHyphenDeletionRecord, citationRelationship, crossReferenceRelationship, reconstruction, requireReconstructionReview, artifact, inspectedArtifact, run, fidelityReceipt, acceptedBaselineSha256, recomputePrivateReconstructionReceiptSha256, rebuildPrivateFidelityReceipt, installCanonicalHyphenDeletionLedger, forgeCanonicalHyphenDeletionLedger, historicalV14Receipt,
+} from './pdf-private-fidelity.cases-support.mjs'
+
+describe('private PDF fidelity boundary cases', () => {
+  it('accepts strict structural boundaries while preserving their separate count', () => {
+    const profiles = ['mobile']
+    const runs = [1, 2].map((ordinal) => {
+      const source = reconstruction()
+      source.lineBoundaryDecisions[0].outcome = 'structural-boundary'
+      source.structurallyConsumedLineBoundaryCount = 1
+      source.completeness.structurallyConsumedLineBoundaryCount = 1
+      return run(ordinal, source, [artifact('mobile', '', source)])
+    })
+    const receipt = createPrivateFidelityReceipt({
+      paperId: 'paper-v1',
+      sourceSha256: hash,
+      byteLength: 123,
+      runs,
+      repeat: 2,
+      profiles,
+    })
+
+    expect(receipt.execution).toMatchObject({
+      lineTransitionGatePassed: true,
+      localValidationPassed: true,
+    })
+    expect(receipt.runs[0].reconstruction).toMatchObject({
+      completeness: {
+        unresolvedCorruptingJoinCount: 0,
+        structurallyConsumedLineBoundaryCount: 1,
+      },
+      lineTransitionEvidence: {
+        decisions: [{ outcome: 'structural-boundary' }],
+      },
+      structure: {
+        unresolvedCorruptingJoinCount: 0,
+        structurallyConsumedLineBoundaryCount: 1,
+      },
+    })
+  })
+
+  it('rejects structural-boundary count mismatches and impossible raw totals', () => {
+    const completenessMismatch = reconstruction()
+    const reportedMismatch = reconstruction()
+    const impossibleRawTotal = reconstruction()
+    for (const source of [
+      completenessMismatch,
+      reportedMismatch,
+      impossibleRawTotal,
+    ]) {
+      source.lineBoundaryDecisions[0].outcome = 'structural-boundary'
+    }
+    completenessMismatch.structurallyConsumedLineBoundaryCount = 1
+    completenessMismatch.completeness.structurallyConsumedLineBoundaryCount = 0
+    reportedMismatch.structurallyConsumedLineBoundaryCount = 0
+    reportedMismatch.completeness.structurallyConsumedLineBoundaryCount = 1
+    impossibleRawTotal.structurallyConsumedLineBoundaryCount = 1
+    impossibleRawTotal.completeness.structurallyConsumedLineBoundaryCount = 1
+    impossibleRawTotal.completeness.unresolvedCorruptingJoinCount = 1
+
+    for (const source of [completenessMismatch, impossibleRawTotal]) {
+      const receipt = createPrivateFidelityReceipt({
+        paperId: 'paper-v1',
+        sourceSha256: hash,
+        byteLength: 123,
+        runs: [1, 2].map((ordinal) =>
+          run(ordinal, source, [artifact('mobile', '', source)]),
+        ),
+        repeat: 2,
+        profiles: ['mobile'],
+      })
+      expect(receipt.execution.lineTransitionGatePassed).toBe(false)
+      expect(receipt.execution.localValidationPassed).toBe(false)
+    }
+    expect(() => artifact('mobile', '', reportedMismatch)).toThrow(
+      'Line transition counts do not match the decision ledger.',
+    )
+  })
+
+  it('fails for review-required or ledger-free reconstruction and changed repeat artifacts', () => {
+    const profiles = ['mobile']
+    const changed = createPrivateFidelityReceipt({
+      paperId: 'paper-v1',
+      sourceSha256: hash,
+      byteLength: 123,
+      runs: [
+        run(1, reconstruction(), [artifact('mobile')]),
+        run(2, reconstruction(), [artifact('mobile', 'f')]),
+      ],
+      repeat: 2,
+      profiles,
+    })
+    const blocked = createPrivateFidelityReceipt({
+      paperId: 'paper-v1',
+      sourceSha256: hash,
+      byteLength: 123,
+      runs: [1, 2].map((ordinal) =>
+        run(ordinal, reconstruction(false), [artifact('mobile')]),
+      ),
+      repeat: 2,
+      profiles,
+    })
+    const missingLedger = createPrivateFidelityReceipt({
+      paperId: 'paper-v1',
+      sourceSha256: hash,
+      byteLength: 123,
+      runs: [1, 2].map((ordinal) =>
+        run(ordinal, reconstruction(true, false), [artifact('mobile')]),
+      ),
+      repeat: 2,
+      profiles,
+    })
+    expect(changed.execution.artifactsDeterministic).toBe(false)
+    expect(changed.passed).toBe(false)
+    expect(blocked.execution.reconstructionDeterministic).toBe(true)
+    expect(blocked.execution.allReady).toBe(false)
+    expect(blocked.passed).toBe(false)
+    expect(missingLedger.execution.lineTransitionGatePassed).toBe(false)
+    expect(missingLedger.passed).toBe(false)
+  })
+
+  it('reports byte-identical artifacts as deterministic without weakening semantic parity', () => {
+    const receipt = fidelityReceipt({
+      transformArtifact(value) {
+        return {
+          ...value,
+          inlineSemanticLedger: {
+            ...value.inlineSemanticLedger,
+            semanticRangeCount:
+              value.inlineSemanticLedger.semanticRangeCount + 1,
+            semanticRangeLedgerSha256: 'f'.repeat(64),
+          },
+        }
+      },
+    })
+
+    expect(receipt.runs[0].artifacts[0]).toMatchObject({
+      structuralValidation: 'passed',
+      receiptSha256: receipt.runs[1].artifacts[0].receiptSha256,
+    })
+    expect(receipt.execution.profileResults[0]).toMatchObject({
+      complete: true,
+      structurallyValid: false,
+      deterministic: true,
+    })
+    expect(receipt.execution.artifactsDeterministic).toBe(true)
+    expect(receipt.execution.localValidationPassed).toBe(false)
+  })
+
+  it('rejects empty, duplicate, non-adjacent, or cross-region transition ledgers', () => {
+    function receiptWithTransitionMutation(mutate) {
+      const profiles = ['mobile']
+      const runs = [1, 2].map((ordinal) => {
+        const source = reconstruction()
+        mutate(source)
+        return run(ordinal, source, [artifact('mobile', '', source)])
+      })
+      return createPrivateFidelityReceipt({
+        paperId: 'paper-v1',
+        sourceSha256: hash,
+        byteLength: 123,
+        runs,
+        repeat: 2,
+        profiles,
+      })
+    }
+
+    const empty = receiptWithTransitionMutation((source) => {
+      source.lineBoundaryDecisions = []
+    })
+    const duplicate = receiptWithTransitionMutation((source) => {
+      source.regions[0].lines.push({ id: 'line-4' })
+      source.lineBoundaryDecisions.push({
+        ...source.lineBoundaryDecisions[0],
+        id: 'transition-2',
+      })
+    })
+    const nonAdjacent = receiptWithTransitionMutation((source) => {
+      source.regions[0].lines.push({ id: 'line-4' })
+      source.lineBoundaryDecisions = [
+        {
+          ...source.lineBoundaryDecisions[0],
+          toLineId: 'line-4',
+        },
+        {
+          ...source.lineBoundaryDecisions[0],
+          id: 'transition-2',
+          fromLineId: 'line-2',
+          toLineId: 'line-4',
+        },
+      ]
+    })
+    const crossRegion = receiptWithTransitionMutation((source) => {
+      source.lineBoundaryDecisions[0].toLineId = 'line-3'
+    })
+
+    for (const receipt of [empty, duplicate, nonAdjacent, crossRegion]) {
+      expect(receipt.execution.lineTransitionGatePassed).toBe(false)
+      expect(receipt.execution.localValidationPassed).toBe(false)
+    }
+  })
+
+  it('keeps transition topology opaque and rejects artifact parity mismatches', () => {
+    const source = reconstruction()
+    source.regions[0].id = 'PRIVATE region prose must not escape'
+    source.regions[0].lines[0].id = 'PRIVATE first line must not escape'
+    source.regions[0].lines[1].id = 'PRIVATE second line must not escape'
+    source.lineBoundaryDecisions[0] = {
+      ...source.lineBoundaryDecisions[0],
+      regionId: source.regions[0].id,
+      fromLineId: source.regions[0].lines[0].id,
+      toLineId: source.regions[0].lines[1].id,
+    }
+    const evidence = createPrivateReconstructionEvidence(source)
+    const serializedTopology = JSON.stringify(evidence.lineTransitionEvidence)
+
+    expect(serializedTopology).not.toContain('PRIVATE')
+    expect(evidence.lineTransitionEvidence.expectedTransitionCount).toBe(1)
+
+    const mismatchKeys = [
+      'canonicalNodeSequenceSha256',
+      'canonicalContentSha256',
+      'relationshipGraphSha256',
+      'assetManifestSha256',
+    ]
+    for (const key of mismatchKeys) {
+      const receipt = fidelityReceipt({
+        transformArtifact(value) {
+          return { ...value, [key]: 'f'.repeat(64) }
+        },
+      })
+      expect(receipt.execution.profileResults[0].structurallyValid).toBe(false)
+      expect(receipt.execution.localValidationPassed).toBe(false)
+    }
+
+    const inlineMismatch = fidelityReceipt({
+      transformArtifact(value) {
+        return {
+          ...value,
+          inlineSemanticLedger: {
+            ...value.inlineSemanticLedger,
+            semanticRangeLedgerSha256: 'f'.repeat(64),
+          },
+        }
+      },
+    })
+    expect(inlineMismatch.execution.profileResults[0].structurallyValid).toBe(
+      false,
+    )
+  })
+
+  it('keeps citation topology opaque and binds anchor-only and target-only changes', () => {
+    const source = reconstruction()
+    source.citationRelationships = [
+      citationRelationship({
+        id: 'PRIVATE citation relationship',
+        referenceRegionId: 'PRIVATE source region',
+        labels: ['PRIVATE citation label'],
+        targetNodeIds: ['PRIVATE bibliography target'],
+        canonicalAnchor: {
+          nodeId: 'PRIVATE canonical anchor',
+          start: 4,
+          end: 7,
+        },
+      }),
+    ]
+    const evidence = createPrivateReconstructionEvidence(source)
+    const serializedCitationGraph = JSON.stringify(
+      evidence.structure.citationRelationshipGraph,
+    )
+    expect(serializedCitationGraph).not.toContain('PRIVATE')
+    expect(serializedCitationGraph).not.toContain('"x"')
+    expect(evidence.structure.citationRelationshipGraph[0].sourceBoxes).toEqual(
+      [expect.stringMatching(/^[a-f0-9]{64}$/)],
+    )
+    expect(evidence.structure).toMatchObject({
+      schemaVersion: '1.7.0',
+      citationRelationshipCount: 1,
+      citationRelationshipCounts: { matched: 1 },
+      citationRelationshipGraphSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
+    })
+
+    const baseline = fidelityReceipt({
+      transformReconstruction(value) {
+        value.citationRelationships = [citationRelationship()]
+        return value
+      },
+    })
+    for (const relationship of [
+      citationRelationship({
+        canonicalAnchor: { nodeId: 'paragraph-private-2', start: 4, end: 7 },
+      }),
+      citationRelationship({ targetNodeIds: ['bibliography-private-2'] }),
+    ]) {
+      const candidate = fidelityReceipt({
+        transformReconstruction(value) {
+          value.citationRelationships = [relationship]
+          return value
+        },
+      })
+      expect(
+        comparePrivateFidelityReceipts(
+          baseline,
+          candidate,
+          acceptedBaselineSha256(baseline),
+        ),
+      ).toEqual({ status: 'failed', passed: false })
+    }
+  })
+
+  it('accepts ambiguous citation relationships in private fidelity evidence', () => {
+    const baseline = fidelityReceipt({
+      transformReconstruction(value) {
+        value.citationRelationships = [
+          citationRelationship({
+            status: 'ambiguous',
+            targetNodeIds: [],
+            candidateNodeIds: [
+              'bibliography-private-1',
+              'bibliography-private-2',
+            ],
+          }),
+        ]
+        return value
+      },
+    })
+
+    expect(baseline.runs[0].reconstruction.structure).toMatchObject({
+      citationRelationshipCount: 1,
+      citationRelationshipCounts: { ambiguous: 1 },
+      citationRelationshipGraph: [
+        expect.objectContaining({
+          candidateNodeIds: [
+            expect.stringMatching(/^[a-f0-9]{64}$/),
+            expect.stringMatching(/^[a-f0-9]{64}$/),
+          ],
+          evidenceSha256s: expect.arrayContaining([
+            expect.stringMatching(/^[a-f0-9]{64}$/),
+          ]),
+          targetNodeIds: [],
+        }),
+      ],
+    })
+    expect(() =>
+      comparePrivateFidelityReceipts(
+        baseline,
+        structuredClone(baseline),
+        acceptedBaselineSha256(baseline),
+      ),
+    ).not.toThrow()
+    const changedCandidates = fidelityReceipt({
+      transformReconstruction(value) {
+        value.citationRelationships = [
+          citationRelationship({
+            status: 'ambiguous',
+            targetNodeIds: [],
+            candidateNodeIds: [
+              'bibliography-private-1',
+              'bibliography-private-3',
+            ],
+          }),
+        ]
+        return value
+      },
+    })
+    const changedEvidence = fidelityReceipt({
+      transformReconstruction(value) {
+        value.citationRelationships = [
+          citationRelationship({
+            status: 'ambiguous',
+            targetNodeIds: [],
+            candidateNodeIds: [
+              'bibliography-private-1',
+              'bibliography-private-2',
+            ],
+            evidence: ['private-citation-evidence-changed'],
+          }),
+        ]
+        return value
+      },
+    })
+    expect(
+      changedCandidates.runs[0].reconstruction.structure
+        .citationRelationshipGraphSha256,
+    ).not.toBe(
+      baseline.runs[0].reconstruction.structure.citationRelationshipGraphSha256,
+    )
+    expect(
+      changedEvidence.runs[0].reconstruction.structure
+        .citationRelationshipGraphSha256,
+    ).not.toBe(
+      baseline.runs[0].reconstruction.structure.citationRelationshipGraphSha256,
+    )
+  })
+
+  it('rejects missing or malformed private citation graph evidence', () => {
+    const baseline = fidelityReceipt({
+      transformReconstruction(value) {
+        value.citationRelationships = [citationRelationship()]
+        return value
+      },
+    })
+    const missing = structuredClone(baseline)
+    delete missing.runs[0].reconstruction.structure.citationRelationshipGraph
+    const malformed = structuredClone(baseline)
+    malformed.runs[0].reconstruction.structure.citationRelationshipGraph[0].canonicalAnchor =
+      { nodeId: '', start: 7, end: 4 }
+    for (const staleRelationship of [
+      citationRelationship({
+        status: 'ambiguous',
+        candidateNodeIds: ['bibliography-private-1', 'bibliography-private-2'],
+      }),
+      citationRelationship({ status: 'unresolved' }),
+    ]) {
+      expect(() =>
+        fidelityReceipt({
+          transformReconstruction(value) {
+            value.citationRelationships = [staleRelationship]
+            return value
+          },
+        }),
+      ).toThrow('Citation relationship receipt state is invalid.')
+    }
+
+    for (const invalid of [missing, malformed]) {
+      expect(() =>
+        comparePrivateFidelityReceipts(
+          invalid,
+          baseline,
+          canonicalJsonHash(invalid),
+        ),
+      ).toThrow('INVALID_PRIVATE_FIDELITY_BASELINE')
+    }
+  })
+
+  it('keeps scholarly cross-reference topology opaque and binds anchor-only and target-only changes', () => {
+    const source = reconstruction()
+    source.crossReferenceRelationships = [
+      crossReferenceRelationship({
+        id: 'PRIVATE cross-reference relationship',
+        text: 'PRIVATE Figure 1',
+        labels: ['PRIVATE Figure 1'],
+        referenceRegionId: 'PRIVATE source region',
+        targets: [
+          {
+            ...crossReferenceRelationship().targets[0],
+            label: 'PRIVATE Figure 1',
+            candidateNodeIds: ['PRIVATE figure target'],
+            targetNodeId: 'PRIVATE figure target',
+          },
+        ],
+        targetNodeIds: ['PRIVATE figure target'],
+        canonicalAnchor: {
+          nodeId: 'PRIVATE canonical anchor',
+          start: 4,
+          end: 12,
+        },
+      }),
+    ]
+    const evidence = createPrivateReconstructionEvidence(source)
+    const serializedGraph = JSON.stringify(
+      evidence.structure.crossReferenceRelationshipGraph,
+    )
+    expect(serializedGraph).not.toContain('PRIVATE')
+    expect(serializedGraph).not.toContain('"x"')
+    expect(
+      evidence.structure.crossReferenceRelationshipGraph[0].sourceBoxes,
+    ).toEqual([expect.stringMatching(/^[a-f0-9]{64}$/)])
+    expect(evidence.structure).toMatchObject({
+      schemaVersion: '1.7.0',
+      crossReferenceRelationshipCount: 1,
+      crossReferenceRelationshipCounts: { 'figure:matched': 1 },
+      crossReferenceRelationshipGraphSha256:
+        expect.stringMatching(/^[a-f0-9]{64}$/),
+    })
+
+    const baseline = fidelityReceipt({
+      transformReconstruction(value) {
+        value.crossReferenceRelationships = [crossReferenceRelationship()]
+        return value
+      },
+    })
+    for (const relationship of [
+      crossReferenceRelationship({
+        canonicalAnchor: {
+          nodeId: 'paragraph-private-2',
+          start: 4,
+          end: 12,
+        },
+      }),
+      crossReferenceRelationship({
+        targets: [
+          {
+            ...crossReferenceRelationship().targets[0],
+            candidateNodeIds: ['figure-private-2'],
+            targetNodeId: 'figure-private-2',
+          },
+        ],
+        targetNodeIds: ['figure-private-2'],
+      }),
+    ]) {
+      const candidate = fidelityReceipt({
+        transformReconstruction(value) {
+          value.crossReferenceRelationships = [relationship]
+          return value
+        },
+      })
+      expect(
+        comparePrivateFidelityReceipts(
+          baseline,
+          candidate,
+          acceptedBaselineSha256(baseline),
+        ),
+      ).toEqual({ status: 'failed', passed: false })
+    }
+  })
+
+  it('rejects missing or malformed private scholarly cross-reference graph evidence', () => {
+    const baseline = fidelityReceipt({
+      transformReconstruction(value) {
+        value.crossReferenceRelationships = [crossReferenceRelationship()]
+        return value
+      },
+    })
+    const missing = structuredClone(baseline)
+    delete missing.runs[0].reconstruction.structure
+      .crossReferenceRelationshipGraph
+    const malformed = structuredClone(baseline)
+    malformed.runs[0].reconstruction.structure.crossReferenceRelationshipGraph[0].targets[0] =
+      {
+        ...malformed.runs[0].reconstruction.structure
+          .crossReferenceRelationshipGraph[0].targets[0],
+        status: 'unresolved',
+        targetNodeId: 'a'.repeat(64),
+      }
+
+    for (const invalid of [missing, malformed]) {
+      expect(() =>
+        comparePrivateFidelityReceipts(
+          invalid,
+          baseline,
+          canonicalJsonHash(invalid),
+        ),
+      ).toThrow('INVALID_PRIVATE_FIDELITY_BASELINE')
+    }
+  })
+
+  it('rejects coordinated artifact and parity tampering against structural evidence', () => {
+    for (const key of [
+      'canonicalContentSha256',
+      'relationshipGraphSha256',
+      'assetManifestSha256',
+    ]) {
+      const source = reconstruction()
+      const first = run(1, source, [artifact('mobile', '', source)])
+      const second = structuredClone(first)
+      second.ordinal = 2
+      for (const candidate of [first, second]) {
+        candidate.reconstruction.artifactParity.publication[key] = 'f'.repeat(
+          64,
+        )
+        candidate.artifacts[0][key] = 'f'.repeat(64)
+        const artifactEvidence = Object.fromEntries(
+          Object.entries(candidate.artifacts[0]).filter(
+            ([field]) => field !== 'receiptSha256',
+          ),
+        )
+        candidate.artifacts[0].receiptSha256 =
+          canonicalJsonHash(artifactEvidence)
+        candidate.reconstructionReceiptSha256 = canonicalJsonHash(
+          candidate.reconstruction,
+        )
+      }
+
+      const receipt = createPrivateFidelityReceipt({
+        paperId: 'paper-v1',
+        sourceSha256: hash,
+        byteLength: 123,
+        runs: [first, second],
+        repeat: 2,
+        profiles: ['mobile'],
+      })
+
+      expect(receipt.execution.profileResults[0].structurallyValid).toBe(false)
+      expect(receipt.execution.localValidationPassed).toBe(false)
+    }
+  })
+
+  it('binds sanitized transition outcomes to the source transition ledger', () => {
+    const runs = [1, 2].map((ordinal) => {
+      const source = reconstruction()
+      const candidate = run(ordinal, source, [artifact('mobile', '', source)])
+      candidate.reconstruction.lineTransitionEvidence.decisions[0].outcome =
+        'no-space'
+      return candidate
+    })
+    const receipt = createPrivateFidelityReceipt({
+      paperId: 'paper-v1',
+      sourceSha256: hash,
+      byteLength: 123,
+      runs,
+      repeat: 2,
+      profiles: ['mobile'],
+    })
+
+    expect(receipt.execution.lineTransitionGatePassed).toBe(false)
+    expect(receipt.execution.localValidationPassed).toBe(false)
+  })
+
+  it('reports an absent frozen baseline explicitly and does not declare acceptance', () => {
+    const profiles = ['mobile']
+    const receipt = createPrivateFidelityReceipt({
+      paperId: 'paper-v1',
+      sourceSha256: hash,
+      byteLength: 123,
+      runs: [1, 2].map((ordinal) =>
+        run(ordinal, reconstruction(), [artifact('mobile')]),
+      ),
+      repeat: 2,
+      profiles,
+    })
+
+    expect(receipt.schemaVersion).toBe('1.9.0')
+    expect(receipt.execution.localValidationPassed).toBe(true)
+    expect(receipt.baselineComparison).toEqual({
+      status: 'not-configured',
+      passed: false,
+    })
+    expect(receipt.passed).toBe(false)
+  })
+
+  it('passes an identical frozen sanitized baseline and ignores timing-only changes', () => {
+    const baseline = fidelityReceipt()
+    const candidate = fidelityReceipt({
+      transformReconstruction(source, ordinal) {
+        source.completeness.readingOrderEvaluation.latencyMs = 912.5 + ordinal
+        source.completeness.readingOrderEvaluation.costUsd = 0.004 * ordinal
+        return source
+      },
+    })
+
+    expect(candidate.execution.reconstructionDeterministic).toBe(true)
+    expect(
+      comparePrivateFidelityReceipts(
+        baseline,
+        candidate,
+        acceptedBaselineSha256(baseline),
+      ),
+    ).toEqual({
+      status: 'passed',
+      passed: true,
+      receiptSha256: canonicalJsonHash(baseline),
+    })
+  })
+
+  it('accepts author-year citation relationships in sanitized evidence', () => {
+    const baseline = fidelityReceipt({
+      transformReconstruction(source) {
+        source.citationRelationships = [
+          citationRelationship({
+            taxonomy: 'author-year-bibliography-citation',
+          }),
+        ]
+        return source
+      },
+    })
+
+    expect(
+      comparePrivateFidelityReceipts(
+        baseline,
+        structuredClone(baseline),
+        acceptedBaselineSha256(baseline),
+      ),
+    ).toMatchObject({ status: 'passed', passed: true })
+  })
+
+  it('requires an independently supplied digest before a not-configured receipt can act as the accepted baseline', () => {
+    const baseline = fidelityReceipt()
+    const candidate = fidelityReceipt()
+
+    expect(baseline.baselineComparison.status).toBe('not-configured')
+    expect(() => comparePrivateFidelityReceipts(baseline, candidate)).toThrow(
+      'INVALID_PRIVATE_FIDELITY_BASELINE',
+    )
+    expect(() =>
+      comparePrivateFidelityReceipts(baseline, candidate, 'f'.repeat(64)),
+    ).toThrow('INVALID_PRIVATE_FIDELITY_BASELINE')
+    expect(
+      comparePrivateFidelityReceipts(
+        baseline,
+        candidate,
+        acceptedBaselineSha256(baseline),
+      ),
+    ).toMatchObject({ status: 'passed', passed: true })
+  })
+
+  it('fails closed for completeness, readiness, structure, diagnostic, or artifact regressions', () => {
+    const baseline = fidelityReceipt()
+    const regressions = [
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.textCoverage = 0.99
+          source.completeness.matchedTextCharacters = 99
+          return source
+        },
+      }),
+      fidelityReceipt({ ready: false }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.paper.nodes.push({
+            id: 'paragraph-regression',
+            source: 'region-regression',
+            type: 'paragraph',
+            text: 'Sanitized structural regression fixture.',
+          })
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.diagnostics.push({
+            severity: 'warning',
+            code: 'NEW_SANITIZED_DIAGNOSTIC',
+          })
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformArtifact(value) {
+          return { ...value, sha256: 'f'.repeat(64) }
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.missingSourceRegionCount = 1
+          return requireReconstructionReview(source, 'MISSING_SOURCE_REGION')
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.unprovenancedRenderedUnitCount = 1
+          return requireReconstructionReview(
+            source,
+            'UNPROVENANCED_RENDERED_UNIT',
+          )
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.mappedInlineSpanCount = 0
+          source.completeness.inlineSpanCoverage = 0
+          return requireReconstructionReview(
+            source,
+            'INCOMPLETE_INLINE_STYLE_COVERAGE',
+          )
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.mappedHyperlinkCount = 0
+          source.completeness.hyperlinkCoverage = 0
+          return requireReconstructionReview(source, 'UNRESOLVED_HYPERLINK')
+        },
+      }),
+    ]
+
+    for (const [index, candidate] of regressions.entries()) {
+      expect(
+        comparePrivateFidelityReceipts(
+          baseline,
+          candidate,
+          acceptedBaselineSha256(baseline),
+        ),
+        `regression fixture ${index}`,
+      ).toEqual({ status: 'failed', passed: false })
+    }
+  })
+
+  it('fails closed when source identity, repeat, or requested profiles differ', () => {
+    const baseline = fidelityReceipt()
+    const mismatches = [
+      fidelityReceipt({ paperId: 'paper-v2' }),
+      fidelityReceipt({ sourceSha256: '9'.repeat(64) }),
+      fidelityReceipt({ byteLength: 124 }),
+      fidelityReceipt({ repeat: 3 }),
+      fidelityReceipt({ profiles: ['paperPro'] }),
+    ]
+
+    for (const candidate of mismatches) {
+      expect(
+        comparePrivateFidelityReceipts(
+          baseline,
+          candidate,
+          acceptedBaselineSha256(baseline),
+        ),
+      ).toEqual({ status: 'failed', passed: false })
+    }
+  })
+
+  it('binds the optional decision-set digest into accepted receipt identity', () => {
+    const acceptedDecisionSetSha256 = 'd'.repeat(64)
+    const changedDecisionSetSha256 = 'e'.repeat(64)
+    const baseline = fidelityReceipt({
+      decisionSetSha256: acceptedDecisionSetSha256,
+    })
+    const matching = fidelityReceipt({
+      decisionSetSha256: acceptedDecisionSetSha256,
+    })
+
+    expect(baseline.decisionSetSha256).toBe(acceptedDecisionSetSha256)
+    expect(fidelityReceipt().decisionSetSha256).toBeNull()
+    expect(
+      comparePrivateFidelityReceipts(
+        baseline,
+        matching,
+        acceptedBaselineSha256(baseline),
+      ),
+    ).toMatchObject({ status: 'passed', passed: true })
+
+    const changed = fidelityReceipt({
+      decisionSetSha256: changedDecisionSetSha256,
+    })
+    expect(
+      comparePrivateFidelityReceipts(
+        baseline,
+        changed,
+        acceptedBaselineSha256(baseline),
+      ),
+    ).toEqual({ status: 'failed', passed: false })
+
+    const missing = structuredClone(matching)
+    delete missing.decisionSetSha256
+    expect(() =>
+      comparePrivateFidelityReceipts(
+        baseline,
+        missing,
+        acceptedBaselineSha256(baseline),
+      ),
+    ).toThrow('INVALID_PRIVATE_FIDELITY_BASELINE')
+
+    const malformed = structuredClone(matching)
+    malformed.decisionSetSha256 = 'not-a-sha256'
+    expect(() =>
+      comparePrivateFidelityReceipts(
+        baseline,
+        malformed,
+        acceptedBaselineSha256(baseline),
+      ),
+    ).toThrow('INVALID_PRIVATE_FIDELITY_BASELINE')
+
+    const forgedBaseline = structuredClone(baseline)
+    forgedBaseline.decisionSetSha256 = changedDecisionSetSha256
+    expect(() =>
+      comparePrivateFidelityReceipts(
+        forgedBaseline,
+        matching,
+        acceptedBaselineSha256(baseline),
+      ),
+    ).toThrow('INVALID_PRIVATE_FIDELITY_BASELINE')
+  })
+
+  it('rejects legacy private receipts that predate complete semantic evidence', () => {
+    const legacy = fidelityReceipt()
+    legacy.schemaVersion = '1.6.0'
+
+    expect(() =>
+      comparePrivateFidelityReceipts(
+        legacy,
+        fidelityReceipt(),
+        acceptedBaselineSha256(legacy),
+      ),
+    ).toThrow('INVALID_PRIVATE_FIDELITY_BASELINE')
+  })
+
+  it('rejects an invalid or locally unaccepted frozen receipt generically', () => {
+    const invalidSchema = { ...fidelityReceipt(), schemaVersion: 'invalid' }
+    const corruptHash = structuredClone(fidelityReceipt())
+    corruptHash.runs[0].reconstructionReceiptSha256 = '0'.repeat(64)
+    const locallyBlocked = fidelityReceipt({ ready: false })
+    const emptyFallbackArtifact = fidelityReceipt({
+      transformArtifact(value) {
+        return {
+          ...value,
+          mode: 'readable-fallback',
+          byteLength: 0,
+          canonicalNodeCount: 0,
+        }
+      },
+    })
+    const invalidInlineCoverage = fidelityReceipt({
+      transformReconstruction(source) {
+        source.completeness.inlineSpanCoverage = 0.5
+        return source
+      },
+    })
+    const invalidCompletenessEvidence = [
+      fidelityReceipt({
+        transformReconstruction(source) {
+          delete source.completeness.expectedHyperlinkCount
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.hyperlinkCoverage = 0.5
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.mappedHyperlinkCount = 2
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.textCoverage = 0.5
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.assetCoverage = 0.5
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.relationshipCoverage = 0.5
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.matchedTextCharacters = 101
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.exportedAssetCount = 2
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.resolvedRelationshipCount = 2
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.decidedLineBoundaryCount = 2
+          return source
+        },
+      }),
+    ]
+
+    for (const baseline of [
+      invalidSchema,
+      corruptHash,
+      locallyBlocked,
+      emptyFallbackArtifact,
+      invalidInlineCoverage,
+      ...invalidCompletenessEvidence,
+    ]) {
+      expect(() =>
+        comparePrivateFidelityReceipts(
+          baseline,
+          fidelityReceipt(),
+          acceptedBaselineSha256(baseline),
+        ),
+      ).toThrow('INVALID_PRIVATE_FIDELITY_BASELINE')
+    }
+  })
+
+  it('rejects private readiness claims that contradict policy, completeness, diagnostics, or blockers', () => {
+    const baseline = fidelityReceipt()
+    const invalidCandidates = [
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.readiness.blockingDiagnosticCodes = ['SOURCE_ERROR']
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.diagnostics.push({
+            severity: 'error',
+            code: 'SOURCE_ERROR',
+          })
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.outputTextCharacters = 97
+          source.completeness.matchedTextCharacters = 97
+          source.completeness.textCoverage = 0.97
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.unresolvedObjectCount = 1
+          source.completeness.unresolvedObjects.assets = 1
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.ocrRequiredPages = [1]
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.completeness.readingOrderDiagnostics = 1
+          source.completeness.readingOrderEvaluation.unresolvedEdgeCount = 1
+          source.completeness.readingOrderEvaluation.reviewRequired = true
+          return source
+        },
+      }),
+      fidelityReceipt({
+        transformReconstruction(source) {
+          source.readiness.ready = false
+          source.readiness.status = 'review-required'
+          return source
+        },
+      }),
+    ]
+
+    for (const candidate of invalidCandidates) {
+      expect(() =>
+        comparePrivateFidelityReceipts(
+          baseline,
+          candidate,
+          acceptedBaselineSha256(baseline),
+        ),
+      ).toThrow('INVALID_PRIVATE_FIDELITY_BASELINE')
+    }
+  })
+
+  it('loads an external baseline without exposing its path or unexpected private fields', async () => {
+    const temporaryDirectory = mkdtempSync(
+      join(tmpdir(), 'pdf-private-baseline-test-'),
+    )
+    const privatePath = join(
+      temporaryDirectory,
+      'operator frozen baseline private path.json',
+    )
+    const privateText = 'PRIVATE SOURCE TEXT MUST NEVER ESCAPE'
+    try {
+      const baseline = fidelityReceipt()
+      const expectedBaselineSha256 = acceptedBaselineSha256(baseline)
+      writeFileSync(privatePath, JSON.stringify(baseline))
+      const comparison = await comparePrivateFidelityBaseline(
+        privatePath,
+        fidelityReceipt(),
+        expectedBaselineSha256,
+      )
+      const serialized = JSON.stringify(comparison)
+      expect(serialized).not.toContain(privatePath)
+      expect(serialized).not.toContain('operator frozen baseline')
+
+      writeFileSync(privatePath, JSON.stringify({ sourceText: privateText }))
+      let failure
+      try {
+        await comparePrivateFidelityBaseline(
+          privatePath,
+          fidelityReceipt(),
+          expectedBaselineSha256,
+        )
+      } catch (error) {
+        failure = String(error)
+      }
+      expect(failure).toBe('Error: INVALID_PRIVATE_FIDELITY_BASELINE')
+      expect(failure).not.toContain(privatePath)
+      expect(failure).not.toContain(privateText)
+    } finally {
+      rmSync(temporaryDirectory, { recursive: true, force: true })
+    }
+  })
+
+  it('requires a new external owner-only directory and exclusive artifact writes', async () => {
+    const temporaryDirectory = mkdtempSync(
+      join(tmpdir(), 'pdf-private-output-test-'),
+    )
+    const output = join(temporaryDirectory, 'new-proof')
+    const linkedOutput = join(temporaryDirectory, 'linked-proof')
+    try {
+      const created = await prepareOwnerOnlyDirectory(output)
+      expect(statSync(created).mode & 0o077).toBe(0)
+      await expect(prepareOwnerOnlyDirectory(created)).rejects.toThrow(
+        'OUTPUT_MUST_BE_NEW_EXTERNAL_DIRECTORY',
+      )
+      symlinkSync(created, linkedOutput)
+      await expect(prepareOwnerOnlyDirectory(linkedOutput)).rejects.toThrow(
+        'OUTPUT_MUST_BE_NEW_EXTERNAL_DIRECTORY',
+      )
+      await expect(prepareOwnerOnlyDirectory('/')).rejects.toThrow(
+        'OUTPUT_MUST_BE_NEW_EXTERNAL_DIRECTORY',
+      )
+      await expect(prepareOwnerOnlyDirectory(homedir())).rejects.toThrow(
+        'OUTPUT_MUST_BE_NEW_EXTERNAL_DIRECTORY',
+      )
+      await expect(
+        prepareOwnerOnlyDirectory(resolve('private-proof-inside-repository')),
+      ).rejects.toThrow('OUTPUT_MUST_BE_NEW_EXTERNAL_DIRECTORY')
+
+      const artifactPath = join(created, 'artifact.epub')
+      await writeExclusive(artifactPath, new Uint8Array([1, 2, 3]))
+      expect(statSync(artifactPath).mode & 0o077).toBe(0)
+      await expect(
+        writeExclusive(artifactPath, new Uint8Array([4, 5, 6])),
+      ).rejects.toMatchObject({ code: 'EEXIST' })
+    } finally {
+      rmSync(temporaryDirectory, { recursive: true, force: true })
+    }
+  })
+
+  it('does not expose an environment-sourced path when execution fails', () => {
+    const temporaryDirectory = mkdtempSync(
+      join(tmpdir(), 'pdf-private-fidelity-test-'),
+    )
+    const privatePath = join(temporaryDirectory, 'operator source.pdf')
+    const outputDirectory = join(temporaryDirectory, 'proof')
+    try {
+      const result = spawnSync(
+        process.execPath,
+        [
+          fileURLToPath(new URL('./pdf-private-fidelity.mjs', import.meta.url)),
+          '--input-env',
+          'SRT_PRIVATE_TEST_PDF',
+          '--paper-id',
+          'paper-v1',
+          '--expected-size',
+          '123',
+          '--expected-sha256',
+          hash,
+          '--profiles',
+          'mobile',
+          '--repeat',
+          '2',
+          '--out',
+          outputDirectory,
+        ],
+        {
+          encoding: 'utf8',
+          env: { ...process.env, SRT_PRIVATE_TEST_PDF: privatePath },
+        },
+      )
+      const output = `${result.stdout}\n${result.stderr}`
+
+      expect(result.status).toBe(2)
+      expect(output).not.toContain(privatePath)
+      expect(output).not.toContain('operator source.pdf')
+      expect(output).toContain(
+        'Private PDF fidelity validation failed without exposing local paths or source content.',
+      )
+    } finally {
+      rmSync(temporaryDirectory, { recursive: true, force: true })
+    }
+  })
+
+  it('emits only a safe error class, code, and message digest when diagnostic output is requested', () => {
+    const temporaryDirectory = mkdtempSync(
+      join(tmpdir(), 'pdf-private-fidelity-diagnostic-test-'),
+    )
+    const privatePath = join(temporaryDirectory, 'missing owner source.pdf')
+    const outputDirectory = join(temporaryDirectory, 'proof')
+    try {
+      const result = spawnSync(
+        process.execPath,
+        [
+          fileURLToPath(new URL('./pdf-private-fidelity.mjs', import.meta.url)),
+          '--input-env',
+          'SRT_PRIVATE_TEST_PDF',
+          '--paper-id',
+          'paper-v1',
+          '--expected-size',
+          '123',
+          '--expected-sha256',
+          hash,
+          '--profiles',
+          'mobile',
+          '--repeat',
+          '2',
+          '--out',
+          outputDirectory,
+          '--safe-error-diagnostic',
+        ],
+        {
+          encoding: 'utf8',
+          env: { ...process.env, SRT_PRIVATE_TEST_PDF: privatePath },
+        },
+      )
+      const output = `${result.stdout}\n${result.stderr}`
+      const diagnosticLine = result.stderr
+        .split('\n')
+        .find((line) => line.startsWith('{'))
+
+      expect(result.status).toBe(2)
+      expect(diagnosticLine).toBeDefined()
+      expect(JSON.parse(diagnosticLine)).toEqual({
+        errorClass: 'Error',
+        errorCode: 'ENOENT',
+        messageSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
+        stage: 'source-stat',
+      })
+      expect(output).not.toContain(privatePath)
+      expect(output).not.toContain('missing owner source.pdf')
+      expect(output).not.toContain('no such file')
+    } finally {
+      rmSync(temporaryDirectory, { recursive: true, force: true })
+    }
+  })
+
+  it('does not reinterpret an arbitrary identifier-shaped error message as a diagnostic code', () => {
+    expect(
+      privateFidelity.safePrivateFailureDiagnostic(
+        new Error('private-source.pdf'),
+        'test-stage',
+      ),
+    ).toEqual({
+      errorClass: 'Error',
+      errorCode: 'UNCLASSIFIED',
+      messageSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
+      stage: 'test-stage',
+    })
+  })
+
+  it('treats EPUBCheck warnings as a pre-publication failure in required mode', () => {
+    const temporaryDirectory = mkdtempSync(
+      join(tmpdir(), 'pdf-private-epubcheck-test-'),
+    )
+    const binaryDirectory = join(temporaryDirectory, 'bin')
+    const argumentsLog = join(temporaryDirectory, 'epubcheck-arguments.jsonl')
+    const outputDirectory = join(temporaryDirectory, 'proof')
+    const inputPath = resolve('tests/fixtures/pdf/born-digital.pdf')
+    const inputBytes = readFileSync(inputPath)
+    const inputSha256 = createHash('sha256').update(inputBytes).digest('hex')
+    try {
+      mkdirSync(binaryDirectory)
+      const fakeEpubCheck = join(binaryDirectory, 'epubcheck')
+      writeFileSync(
+        fakeEpubCheck,
+        `#!/usr/bin/env node
+const { appendFileSync } = require('node:fs')
+if (process.argv.includes('--version')) process.exit(0)
+appendFileSync(process.env.EPUBCHECK_ARGUMENTS_LOG, JSON.stringify(process.argv.slice(2)) + '\\n')
+process.exit(1)
+`,
+      )
+      chmodSync(fakeEpubCheck, 0o755)
+
+      const result = spawnSync(
+        process.execPath,
+        [
+          fileURLToPath(new URL('./pdf-private-fidelity.mjs', import.meta.url)),
+          '--input-env',
+          'SRT_PRIVATE_TEST_PDF',
+          '--paper-id',
+          'public-fixture',
+          '--expected-size',
+          String(inputBytes.byteLength),
+          '--expected-sha256',
+          inputSha256,
+          '--profiles',
+          'mobile',
+          '--repeat',
+          '2',
+          '--out',
+          outputDirectory,
+          '--require-epubcheck',
+          '--safe-error-diagnostic',
+        ],
+        {
+          encoding: 'utf8',
+          env: {
+            ...process.env,
+            EPUBCHECK_ARGUMENTS_LOG: argumentsLog,
+            PATH: `${binaryDirectory}:${process.env.PATH ?? ''}`,
+            SRT_PRIVATE_TEST_PDF: inputPath,
+          },
+          timeout: 120_000,
+        },
+      )
+      const diagnostic = JSON.parse(
+        result.stderr.split('\n').find((line) => line.startsWith('{')),
+      )
+
+      expect(result.status, result.stderr).toBe(2)
+      expect(diagnostic).toMatchObject({
+        errorClass: 'Error',
+        errorCode: 'EPUBCHECK_FAILED',
+        stage: 'epubcheck-validate',
+      })
+      expect(
+        readFileSync(argumentsLog, 'utf8')
+          .trim()
+          .split('\n')
+          .map((line) => JSON.parse(line)),
+      ).toEqual([
+        ['--failonwarnings', expect.stringMatching(/publication\.epub$/)],
+      ])
+      expect(readdirSync(outputDirectory)).toEqual([])
+      expect(`${result.stdout}\n${result.stderr}`).not.toContain(inputPath)
+    } finally {
+      rmSync(temporaryDirectory, { recursive: true, force: true })
+    }
+  }, 120_000)
+
+  it('records every required EPUBCheck pass in the sanitized private receipt', () => {
+    const temporaryDirectory = mkdtempSync(
+      join(tmpdir(), 'pdf-private-epubcheck-pass-test-'),
+    )
+    const binaryDirectory = join(temporaryDirectory, 'bin')
+    const argumentsLog = join(temporaryDirectory, 'epubcheck-arguments.jsonl')
+    const outputDirectory = join(temporaryDirectory, 'proof')
+    const inputPath = resolve('tests/fixtures/pdf/born-digital.pdf')
+    const inputBytes = readFileSync(inputPath)
+    const inputSha256 = createHash('sha256').update(inputBytes).digest('hex')
+    try {
+      mkdirSync(binaryDirectory)
+      const fakeEpubCheck = join(binaryDirectory, 'epubcheck')
+      writeFileSync(
+        fakeEpubCheck,
+        `#!/usr/bin/env node
+const { appendFileSync } = require('node:fs')
+if (process.argv.includes('--version')) process.exit(0)
+appendFileSync(process.env.EPUBCHECK_ARGUMENTS_LOG, JSON.stringify(process.argv.slice(2)) + '\\n')
+`,
+      )
+      chmodSync(fakeEpubCheck, 0o755)
+
+      const result = spawnSync(
+        process.execPath,
+        [
+          fileURLToPath(new URL('./pdf-private-fidelity.mjs', import.meta.url)),
+          '--input-env',
+          'SRT_PRIVATE_TEST_PDF',
+          '--paper-id',
+          'public-fixture',
+          '--expected-size',
+          String(inputBytes.byteLength),
+          '--expected-sha256',
+          inputSha256,
+          '--profiles',
+          'mobile',
+          '--repeat',
+          '2',
+          '--out',
+          outputDirectory,
+          '--require-epubcheck',
+        ],
+        {
+          encoding: 'utf8',
+          env: {
+            ...process.env,
+            EPUBCHECK_ARGUMENTS_LOG: argumentsLog,
+            PATH: `${binaryDirectory}:${process.env.PATH ?? ''}`,
+            SRT_PRIVATE_TEST_PDF: inputPath,
+          },
+          timeout: 120_000,
+        },
+      )
+      const receipt = JSON.parse(result.stdout)
+
+      expect(result.status, result.stderr).toBe(1)
+      expect(receipt).toMatchObject({
+        schemaVersion: '1.9.0',
+        execution: {
+          epubCheckRequired: true,
+          epubCheckPassedCount: 2,
+          allEpubCheckPassed: true,
+        },
+      })
+      expect(
+        receipt.runs
+          .flatMap((run) => run.artifacts)
+          .map((artifact) => artifact.epubCheck),
+      ).toEqual([{ status: 'passed' }, { status: 'passed' }])
+      expect(readdirSync(outputDirectory).sort()).toEqual([
+        'private-fidelity-receipt.json',
+        'run-1-mobile.epub',
+        'run-2-mobile.epub',
+      ])
+      expect(
+        readFileSync(argumentsLog, 'utf8').trim().split('\n'),
+      ).toHaveLength(2)
+    } finally {
+      rmSync(temporaryDirectory, { recursive: true, force: true })
+    }
+  }, 120_000)
+
+  it('replays a hash-pinned empty decision set without exposing either local path', () => {
+    const temporaryDirectory = mkdtempSync(
+      join(tmpdir(), 'pdf-private-decisions-test-'),
+    )
+    const inputPath = resolve('tests/fixtures/pdf/born-digital.pdf')
+    const inputBytes = readFileSync(inputPath)
+    const inputSha256 = createHash('sha256').update(inputBytes).digest('hex')
+    const decisionPath = join(temporaryDirectory, 'owner decisions.json')
+    const decisionBytes = Buffer.from(
+      `${JSON.stringify({
+        schemaVersion: '1.1.0',
+        documentSha256: inputSha256,
+        decisions: [],
+      })}\n`,
+    )
+    const decisionSha256 = createHash('sha256')
+      .update(decisionBytes)
+      .digest('hex')
+    const outputDirectory = join(temporaryDirectory, 'proof')
+    try {
+      writeFileSync(decisionPath, decisionBytes, { mode: 0o600 })
+      const result = spawnSync(
+        process.execPath,
+        [
+          fileURLToPath(new URL('./pdf-private-fidelity.mjs', import.meta.url)),
+          '--input-env',
+          'SRT_PRIVATE_TEST_PDF',
+          '--paper-id',
+          'public-fixture',
+          '--expected-size',
+          String(inputBytes.byteLength),
+          '--expected-sha256',
+          inputSha256,
+          '--profiles',
+          'mobile',
+          '--repeat',
+          '2',
+          '--out',
+          outputDirectory,
+          '--decisions-env',
+          'SRT_PRIVATE_TEST_DECISIONS',
+          '--expected-decisions-sha256',
+          decisionSha256,
+        ],
+        {
+          encoding: 'utf8',
+          env: {
+            ...process.env,
+            SRT_PRIVATE_TEST_PDF: inputPath,
+            SRT_PRIVATE_TEST_DECISIONS: decisionPath,
+          },
+          timeout: 120_000,
+        },
+      )
+      const output = `${result.stdout}\n${result.stderr}`
+
+      expect(result.status, result.stderr).toBe(1)
+      expect(JSON.parse(result.stdout)).toMatchObject({
+        decisionSetSha256: decisionSha256,
+        execution: {
+          reconstructionDeterministic: true,
+          artifactsDeterministic: true,
+          localValidationPassed: true,
+        },
+        baselineComparison: { status: 'not-configured', passed: false },
+      })
+      expect(output).not.toContain(inputPath)
+      expect(output).not.toContain(decisionPath)
+      expect(output).not.toContain('owner decisions.json')
+    } finally {
+      rmSync(temporaryDirectory, { recursive: true, force: true })
+    }
+  }, 120_000)
+
+})
